WHAM!
Mapping works!!

it was a pain to have to go move the centre point using translation but once it was done I could map a scale to one of the audio tracks. its working how I want so now I just have to figure out the next bit to be mapped. I'm playing around with mapping colours which is certainly interesting